+++ src/main/java/no/nav/tag/tiltaksgjennomforing/avtale/EndreAvtale.java
@@ -30,7 +30,7 @@ public class EndreAvtale {
 
     private LocalDate startDato;
     private LocalDate sluttDato;
-    private Integer stillingprosent;
+    private Double stillingprosent;
     private String arbeidsoppgaver;
     private String stillingstittel;
     private Integer stillingStyrk08;
