+++ src/main/java/no/nav/tag/tiltaksgjennomforing/avtale/EndreAvtaleArena.java
@@ -34,7 +34,7 @@ public class EndreAvtaleArena implements Comparable<Avtale> {
 
     private LocalDate startdato;
     private LocalDate sluttdato;
-    private Integer stillingprosent;
+    private Double stillingprosent;
     private Double antallDagerPerUke;
     private Handling handling;
